@@ -1,67 +1,43 @@
-import React from 'react'; 
-import HomePage from './pages/homepage/homepage.component';
-import { Switch, Route, Redirect } from 'react-router-dom';
-import { connect } from 'react-redux';
-import ShopPage from './pages/shoppage/shoppage.component';
-import Header from './components/header/header.component';
-import SignInAndSignUpPage from './pages/sign-in-sign-uppage/sign-in-sign-up.component';
-import { auth, createUserProfileDocument } from './firebase/firebase.utils';
-import { setCurrentUser } from './redux/user/user.actions';
-import { selectCurrentUser } from './redux/user/user.selectors';
-import { createStructuredSelector } from 'reselect';
-import CheckoutPage from './pages/checkoutpage/checkoutpage.component';
-import './App.css';
+const App = () => {
+  return (
+    <div className="categories-container">
+      <div className="category-container">
+        {/* <img /> */}
+        <div className="category-body-container">
+          <h2>Hats</h2>
+          <p>Shop Now</p>
+        </div>
+      </div>
+      <div className="category-container">
+        {/* <img /> */}
+        <div className="category-body-container">
+          <h2>Jackets</h2>
+          <p>Shop Now</p>
+        </div>
+      </div>
+      <div className="category-container">
+        {/* <img /> */}
+        <div className="category-body-container">
+          <h2>Sneakers</h2>
+          <p>Shop Now</p>
+        </div>
+      </div>
+      <div className="category-container">
+        {/* <img /> */}
+        <div className="category-body-container">
+          <h2>Womens</h2>
+          <p>Shop Now</p>
+        </div>
+      </div>
+      <div className="category-container">
+        {/* <img /> */}
+        <div className="category-body-container">
+          <h2>Mens</h2>
+          <p>Shop Now</p>
+        </div>
+      </div>
+    </div>
+  );
+};
 
-class App extends React.Component {
-
-    unsubscribeFromAuth = null;
-
-    componentDidMount() {
-        const {setCurrentUser} = this.props;
-
-        this.unsubscribeFromAuth = auth.onAuthStateChanged(async userAuth => {
-            if (userAuth){
-                const userRef = await createUserProfileDocument(userAuth);
-
-                userRef.onSnapshot(snapShot => {
-                    setCurrentUser({
-                        id: snapShot.id,
-                        ...snapShot.data()
-                    })
-                });
-            } else {
-                setCurrentUser(userAuth);
-            }
-        });
-    }
-
-    componentWillUnmount() {
-        this.unsubscribeFromAuth();
-    }
-
-    render(){
-        return (
-            <div>
-                <Header/>
-                <Switch>
-                    <Route exact path='/' component={HomePage}/>
-                    <Route path='/shop' component={ShopPage}/>
-                    <Route exact path='/checkout' component={CheckoutPage}/>
-                    <Route exact path='/signin' render={() => this.props.currentUser ? (<Redirect to='/'/>) : (<SignInAndSignUpPage/>)}/>
-                </Switch>
-            </div>
-        );
-    }
-}
-
-// This allows the component to use the state
-const mapStateToProps = createStructuredSelector({
-    currentUser: selectCurrentUser
-});
-
-// this sets the state in redux
-const mapDispatchToProps = dispatch => ({
-    setCurrentUser: user => dispatch(setCurrentUser(user))
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
